\name{runbLS}
\alias{runbLS}

\title{
Run bLSmodelR
}
\description{
Run \sQuote{bLSmodelR} using the specified model input.
}
\usage{
runbLS(ModelInput, Cat.Path = NULL, ncores = NULL, TDonly = NULL, 
  asDT = TRUE, simpleNames = asDT, memory_limit = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ModelInput}{
a \code{list} of data.frames defining the model input as generated by \code{\link{genInputList}}.
}
  \item{Cat.Path}{
optional. Character string specifying the path to the touchdown catalogs folder.
}
  \item{ncores}{
integer. Number of cores used for parallel processing.
}
  \item{TDonly}{
logical. Skip calculation of C/E etc., only simulate touchdown catalogs? Default is \code{FALSE}.
}
  \item{asDT}{
logical. If \code{TRUE}, the Output will be a \code{data.table}. The default \code{FALSE} returns a \code{data.frame}.
}
  \item{simpleNames}{
logical. Use simple column names for Output, if \code{TRUE}. Default \code{FALSE} provides more descriptive column names.
}
  \item{memory_limit}{
character. Maximal vector heap size (memory limit) specified via 'R_MAX_VSIZE' (see also \code{?Memory}). This has only an effect if R is running in a bash environment.
}
}
\details{
	It is possible to integrate the entire function in parallel calculation, i.e. on different nodes. The results from different calls can afterwards be merged using the function \code{\link{join.bLSresult}}. 
}
\value{
A \code{data.frame} (or \code{data.table}) of class \code{bLSresult} containing the results of the model run.
The following \code{attributes} are additionally attached:
\describe{
\item{CalcSteps}{a \code{data.table} containing the individual model calculation intervals. Keys: \code{rn,Sensor}}
\item{CatPath}{path to catalog folder}
\item{Catalogs}{a \code{data.table} containing the catalog names. Keys: \code{rn,Sensor,Cat.Name}}
\item{ModelInput}{the ModelInput as supplied.}
}

}

\author{
Christoph Haeni
}
% \note{
% %%
% }

%% ~Make other sections like Warning with \section{Warning }{....} ~
\references{
Flesch, T. K., J. D. Wilson, et al. (1995). \dQuote{Backward-time Lagrangian stochastic dispersion models and their application to estimate gaseous emissions.} Journal of Applied Meteorology 34(6): 1320-1332.
}

\seealso{
\code{\link{bLSmodelR-package}}, \code{\link{genInputList}}, \code{\link{coreModel}}, \code{\link{join.bLSresult}}.
}
\examples{
\dontrun{
# Example Run with default values:
Sensor <- genSensors(PointSensor=list(x=0,y=0,z=2))
Ints <- genInterval(N0=1E4)
InputList <- genInputList(Sensor,Ints)

## run model
Run <- runbLS(InputList,Cat.Path=getwd())

## get catalog
TDcat <- readCatalog(getCatalogs(Run))

# look at TD catalog:
TDcat
plot(TDcat)

}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
